Update router.js
I'd ask we don't commit changes to the router when they're made for testing - keep those local and keep router on dev branch in a clean state. When you're done testing, do a git checkout development -- ./src/router.js.  
If you need to access a different version you can check it out from another branch or previous commit but the more this file is in flux and we're adding routes and saving them, the more cleanup we have to do later.