Fixed problem that data wasn't appearing when module page first loaded
Needed to go back to the style of returning immediately from Module1/2/3.js if still loading.  If I allow the <Module> to be loaded it will load its children even though it wasn't dsplaying them b/c the system was still loading.  The problem is the children initialize thier local state from the store when instantiated.  So we have to guard against instatiation until the Store is loaded.